Refactored model stub by shifiting it to test util. Added piechart.css
