+++ src/main/resources/view/PieChart.css
@@ -1,0 +1,3 @@
+.chart-title {
+    -fx-font-size: 2em;
+}
