+++ src/test/java/seedu/address/logic/commands/MergePersonCommandTest.java
@@ -28,6 +28,7 @@ import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.policy.Policy;
 import seedu.address.model.policy.PolicyName;
+import seedu.address.testutil.TestUtil.ModelStub;
 
 public class MergePersonCommandTest {
 
@@ -80,142 +81,6 @@ public class MergePersonCommandTest {
 
         // different person -> returns false
         assertFalse(commandWithAlice.equals(commandWithBob));
-    }
-
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Person getPerson(Person person) {
-            return person;
-        }
-
-        @Override
-        public void addPolicy(Policy policy) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Policy getPolicy(Policy policy) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Policy getPolicyWithName(PolicyName policyName) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPolicy(Policy policy) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPolicyWithName(PolicyName policyName) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePolicy(Policy target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPolicy(Policy target, Policy editedPolicy) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Policy> getFilteredPolicyList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableMap<String, Integer> getPolicyPopularityBreakdown() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableMap<String, Integer> getAgeGroupBreakdown() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPolicyList(Predicate<Policy> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 
     /**
